+++ src/generate/arr_rand.cpp
@@ -72,17 +72,17 @@ err_type exe_gen_arr_rand(const input_list& inputs, ifstream& fin) {
 		return err_type::test_format;
 	}
 
-	uint32_t n, ntrees, nit;
+	uint64_t n, ntrees, nit;
 	while (fin >> n >> ntrees >> nit) {
 		// do 'ntrees' trees of 'n' vertices
 		rand_ulab_free_trees TreeGen(n);
 
-		for (uint32_t nt = 0; nt < ntrees; ++nt) {
+		for (uint64_t nt = 0; nt < ntrees; ++nt) {
 			const free_tree T = TreeGen.get_tree();
 
 			rand_arrangements RandArr(T, 100);
 
-			for (uint32_t it = 0; it < nit; ++it) {
+			for (uint64_t it = 0; it < nit; ++it) {
 				const linear_arrangement arr = RandArr.get_arrangement();
 
 				// Do some sanity checks.
@@ -97,7 +97,7 @@ err_type exe_gen_arr_rand(const input_list& inputs, ifstream& fin) {
 				}
 			}
 
-			for (uint32_t it = 0; it < nit; ++it) {
+			for (uint64_t it = 0; it < nit; ++it) {
 				const linear_arrangement arr = RandArr.yield_arrangement();
 
 				// Do some sanity checks.
