+++ src/graphs/construct_utils.cpp
@@ -53,7 +53,7 @@ namespace tests {
 namespace graphs {
 
 vector<edge> enumerate_edges_brute_force(const directed_graph& g) {
-	const uint32_t n = g.get_num_nodes();
+	const uint64_t n = g.get_num_nodes();
 	set<edge> E;
 	for (node s = 0; s < n; ++s) {
 	for (auto t : g.get_out_neighbours(s)) {
@@ -67,7 +67,7 @@ vector<edge> enumerate_edges_brute_force(const directed_graph& g) {
 }
 
 vector<edge> enumerate_edges_brute_force(const undirected_graph& g) {
-	const uint32_t n = g.get_num_nodes();
+	const uint64_t n = g.get_num_nodes();
 	set<edge> E;
 	for (node s = 0; s < n; ++s) {
 	for (auto t : g.get_neighbours(s)) {
@@ -91,7 +91,7 @@ bool share_vertices(const edge_pair& st_uv) {
 }
 
 vector<edge_pair> enumerate_Q_brute_force(const undirected_graph& g) {
-	const uint32_t n = g.get_num_nodes();
+	const uint64_t n = g.get_num_nodes();
 	set<edge_pair> Q;
 	for (node s = 0; s < n; ++s) {
 	for (node t : g.get_neighbours(s)) {
@@ -121,7 +121,7 @@ vector<edge_pair> enumerate_Q_brute_force(const undirected_graph& g) {
 }
 
 vector<edge_pair> enumerate_Q_brute_force(const directed_graph& g) {
-	const uint32_t n = g.get_num_nodes();
+	const uint64_t n = g.get_num_nodes();
 	set<edge_pair> Q;
 	for (node s = 0; s < n; ++s) {
 	for (node t : g.get_out_neighbours(s)) {
